cpp files for common_search
still non-functional
